requirements.txt
pypdf2==3.0.1
openai==0.27.8
python-dotenv==1.0.0
